Exit if sync fails
This allows to automate capturing with ccch_scan. Sometimes after dumping a burst the program gets stuck with this sync error
and never exits. You probably want to restart capturing in this case, and this fix allows that, by simply exiting, and letting your automatization tool of choice written in some script language continue in its while loop and restart ccch_scan. If you write such a script, it's a good idea to wait a few secs (5?) before restart. If you aren't automating anything, this gives you a clear sign of an error, and you should restart ccch_scan manually.